Move `memo` to the `schema` package
