+++ internal/transactionsubmission/store/transactions_test.go
@@ -15,12 +15,12 @@ import (
 
 	"github.com/stellar/stellar-disbursement-platform-backend/db"
 	"github.com/stellar/stellar-disbursement-platform-backend/db/dbtest"
-	"github.com/stellar/stellar-disbursement-platform-backend/internal/transactionsubmission/utils"
+	"github.com/stellar/stellar-disbursement-platform-backend/pkg/schema"
 )
 
 func Test_Transaction_BuildMemo(t *testing.T) {
 	testCases := []struct {
-		memoType        utils.MemoType
+		memoType        schema.MemoType
 		memoValue       string
 		wantMemo        txnbuild.Memo
 		wantErrContains string
@@ -31,45 +31,45 @@ func Test_Transaction_BuildMemo(t *testing.T) {
 			wantMemo:  nil,
 		},
 		{
-			memoType:        utils.MemoTypeText,
+			memoType:        schema.MemoTypeText,
 			memoValue:       "This is a very long text that should exceed the 28-byte limit",
 			wantErrContains: "text memo must be 28 bytes or less",
 		},
 		{
-			memoType:        utils.MemoTypeText,
+			memoType:        schema.MemoTypeText,
 			memoValue:       "HelloWorld!",
 			wantMemo:        txnbuild.MemoText("HelloWorld!"),
 			wantErrContains: "",
 		},
 		{
-			memoType:        utils.MemoTypeID,
+			memoType:        schema.MemoTypeID,
 			memoValue:       "not-a-valid-uint64",
 			wantErrContains: "invalid Memo ID value, must be a uint64",
 		},
 		{
-			memoType:        utils.MemoTypeID,
+			memoType:        schema.MemoTypeID,
 			memoValue:       "1234567890",
 			wantMemo:        txnbuild.MemoID(1234567890),
 			wantErrContains: "",
 		},
 		{
-			memoType:        utils.MemoTypeHash,
+			memoType:        schema.MemoTypeHash,
 			memoValue:       "12f37f82eb6708daa0ac372a1a67a0f33efa6a9cd213ed430517e45fefb5157712f37f82eb6708daa0ac372a1a67a0f33efa6a9cd213ed430517e45fefb51577",
 			wantErrContains: "hash memo must be 64 hex characters (32 bytes)",
 		},
 		{
-			memoType:        utils.MemoTypeHash,
+			memoType:        schema.MemoTypeHash,
 			memoValue:       "12f37f82eb6708daa0ac372a1a67a0f33efa6a9cd213ed430517e45fefb51577",
 			wantMemo:        txnbuild.MemoHash([]byte{0x12, 0xf3, 0x7f, 0x82, 0xeb, 0x67, 0x08, 0xda, 0xa0, 0xac, 0x37, 0x2a, 0x1a, 0x67, 0xa0, 0xf3, 0x3e, 0xfa, 0x6a, 0x9c, 0xd2, 0x13, 0xed, 0x43, 0x05, 0x17, 0xe4, 0x5f, 0xef, 0xb5, 0x15, 0x77}),
 			wantErrContains: "",
 		},
 		{
-			memoType:        utils.MemoTypeReturn,
+			memoType:        schema.MemoTypeReturn,
 			memoValue:       "12f37f82eb6708daa0ac372a1a67a0f33efa6a9cd213ed430517e45fefb5157712f37f82eb6708daa0ac372a1a67a0f33efa6a9cd213ed430517e45fefb51577",
 			wantErrContains: "return memo must be 64 hex characters (32 bytes)",
 		},
 		{
-			memoType:        utils.MemoTypeReturn,
+			memoType:        schema.MemoTypeReturn,
 			memoValue:       "12f37f82eb6708daa0ac372a1a67a0f33efa6a9cd213ed430517e45fefb51577",
 			wantMemo:        txnbuild.MemoReturn([]byte{0x12, 0xf3, 0x7f, 0x82, 0xeb, 0x67, 0x08, 0xda, 0xa0, 0xac, 0x37, 0x2a, 0x1a, 0x67, 0xa0, 0xf3, 0x3e, 0xfa, 0x6a, 0x9c, 0xd2, 0x13, 0xed, 0x43, 0x05, 0x17, 0xe4, 0x5f, 0xef, 0xb5, 0x15, 0x77}),
 			wantErrContains: "",
