+++ pkg/schema/memo.go
@@ -1,4 +1,4 @@
-package utils
+package schema
 
 import (
 	"encoding/hex"
