+++ pkg/schema/memo_test.go
@@ -1,4 +1,4 @@
-package utils
+package schema
 
 import (
 	"fmt"
